(WIP) initial
This commit only includes the bare minimum to POST a json file to the Kotlin backend and then gets serialized. currently the json gets directly handed over by the backend to the frontend. The backend will organize the layout of the frontend and provide access to Kotlin apis for the LLM etc...